Removed bootstrap file and injected axios code directly to app.js
