+++ resources/assets/js/app.js
@@ -1,18 +1,30 @@
-// External Setup
-require('./axios');
+/**
+ * Setup basic Axios AJAX headers and settings
+ */
+import axios from 'axios';
+window.axios = axios;
+window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
+
+// Setup CSRF token
+let token = document.head.querySelector('meta[name="csrf-token"]');
+
+if (token) {
+    window.axios.defaults.headers.common['X-CSRF-TOKEN'] = token.content;
+} else {
+    console.error('CSRF token not found: https://laravel.com/docs/csrf#csrf-x-csrf-token');
+}
+
 // require('./echo');
 
-window.Vue = require('vue');
-
 /**
- * Next, we will create a fresh Vue application instance and attach it to
- * the page. Then, you may begin adding components to this application
- * or customize the JavaScript scaffolding to fit your unique needs.
+ * Setup Vue for the project
  */
+import Vue from 'vue';
+window.Vue = Vue;
 
+// Vue Comonent
 const FacetLandingPage = Vue.component('facet-landing-page', require('./components/FacetLandingPage.vue'));
 
 const app = new Vue({
-    el: '#app',
     component: { FacetLandingPage }
-});
+}).$mount('#app');
